Remove function level import and fix circular import
