+++ mlrun/runtimes/base.py
@@ -30,6 +30,7 @@ import mlrun.common.runtimes
 import mlrun.common.schemas
 import mlrun.common.schemas.model_monitoring.constants as mm_constants
 import mlrun.errors
+import mlrun.runtimes.constants
 import mlrun.launcher.factory
 import mlrun.utils.helpers
 import mlrun.utils.notifications
@@ -446,12 +447,9 @@ class BaseRuntime(ModelObj):
             "MLRUN_DEFAULT_PROJECT": active_project,
         }
 
-        # Import here to avoid circular import
-        import mlrun.runtimes
-
         # Set auth session only for nuclio runtimes that have an access key
         if (
-            self.kind in mlrun.runtimes.RuntimeKinds.nuclio_runtimes()
+            self.kind in mlrun.runtimes.constants.RuntimeKinds.nuclio_runtimes()
             and self.metadata.credentials.access_key
         ):
             runtime_env[
